Updated basicInstallation.sh file.
Former-commit-id: 21db2358db093b0293abdab63225a72349b142af

+++ scripts/vm-initial-setup/basicInstallations.sh
@@ -3,19 +3,27 @@
 
 #echo "Installing APT"
 #wget http://pkgs.repoforge.org/rpmforge-release/rpmforge-release-0.5.3-1.el6.rf.x86_64.rpm
-
 #rpm -i rpmforge-release-0.5.*.rpm
-
 #yum install apt
-
 #sudo apt-get update
 
-echo "Installing Net tools (ifconfig)"
-yum install -y net-tools
+echo ...........................Installing Net tools (ifconfig)...........................
+if [ -f "/vshare/base-images/misc/net-tools-2.0-0.17.20131004git.el7.x86_64.rpm" ]; then
+	yum install -y /vshare/base-images/misc/net-tools-2.0-0.17.20131004git.el7.x86_64.rpm
+else
+	yum install -y net-tools
+fi
 
-echo "Installing telnet"
-yum install -y telnet
+echo ...........................Installing telnet...........................
+if [ -f "/vshare/base-images/misc/telnet-0.17-59.el7.x86_64.rpm" ]; then
+	yum install -y /vshare/base-images/misc/telnet-0.17-59.el7.x86_64.rpm
+else
+	yum install -y telnet
+fi
 
-
-echo "Installing wget"
-yum install -y wget
+echo ...........................Installing wget...........................
+if [ -f "/vshare/base-images/misc/wget-1.14-10.el7.x86_64.rpm" ]; then
+	yum install -y /vshare/base-images/misc/wget-1.14-10.el7.x86_64.rpm
+else
+	yum install -y wget
+fi
